+++ src/main/java/com/cobber/fta/LogicalTypeCode.java
@@ -21,8 +21,7 @@ import java.util.Locale;
 import com.cobber.fta.core.FTAPluginException;
 
 /**
- * All Logical Types that support the {@link LTRandom} interface should extend this abstract class.  This interface
- * provides a {@link LTRandom#nextRandom} which will create a new valid example of the Semantic Type.
+ * All Logical Types implemented via Java code typically extend this class.
  */
 public abstract class LogicalTypeCode extends LogicalType {
 	protected SecureRandom random;