+++ src/main/java/com/cobber/fta/LogicalTypeRegExp.java
@@ -28,6 +28,9 @@ import com.cobber.fta.token.TokenStreams;
 
 import nl.flotsam.xeger.Xeger;
 
+/**
+ * This class supports all plugins that are based on a RegExp.
+ */
 public class LogicalTypeRegExp extends LogicalType {
 	private static final String WRONG_TYPE = "LogicalTypeRegExp baseType must be LONG, DOUBLE or STRING, not ";
 	private Pattern pattern;